separate component for HeaderAuth
