+++ src/components/Header/header.module.scss
@@ -35,45 +35,6 @@
         width: 100px;
         height: 80px;
     }
-
-    &__button_container{
-        display: flex;
-        align-items: center;
-        justify-content: flex-start;
-
-        gap: 15px;
-    }
-
-    &__auth_container{
-        display: flex;
-        align-items:center;
-        justify-content: flex-start;
-        gap: 20px;
-    }
-
-    &__data_container{
-        display: flex;
-        align-items:center;
-        justify-content: flex-start;
-        gap: 5px;
-    }
-
-    &__user_img{
-        filter: brightness(0) saturate(100%) invert(99%) sepia(2%) saturate(423%) hue-rotate(216deg) brightness(117%) contrast(100%);
-    }
-
-    &__favorites_img{
-        width: 35px;
-        height: 35px;
-        cursor: pointer;
-    }
-
-    &__login_text{
-        font-size: 16px;
-        font-weight: 400;
-        line-height: 120%;
-        cursor: default;
-    }
 }
 
 
@@ -85,10 +46,6 @@
 @media only screen and (max-width: 800px) {
     .header{ 
         padding: 10px 20px; 
-
-        &__auth_container{
-            display: none;
-        }
 
         &__logo_img{
             width: 80px;
